09/requirements.txt
langchain==0.1.15
langgraph==0.0.32
tavily-python==0.3.3
openai
langchain-openai
python-dotenv
retry
